@@ -4,10 +4,8 @@ namespace Parable\Di\Tests\Classes;
 
 class FakeWithInterfaceDependency
 {
-    public $fakeInterfaceObject;
-
-    public function __construct(FakeInterface $fakeInterfaceObject)
-    {
-        $this->fakeInterfaceObject = $fakeInterfaceObject;
+    public function __construct(
+        public FakeInterface $fakeInterfaceObject
+    ) {
     }
 }
