@@ -8,6 +8,9 @@ class NotFoundException extends Exception
 {
     public static function fromId(string $id)
     {
-        return new self(sprintf("No instance found stored for `%s`.", $id));
+        return new self(sprintf(
+            "No instance found stored for `%s`.",
+            $id
+        ));
     }
 }
